+++ src/value.rs
@@ -373,6 +373,7 @@ pub struct TypedArrayData {
 
 pub struct ArrayBufferData {
     pub bytes: Mutex<Vec<u8>>,
+    pub detached: AtomicBool,
     pub props: Mutex<IndexMap<PropertyKey, PropertyDescriptor>>,
     pub proto: Mutex<Option<Value>>,
 }